Move infraction manage and utility

+++ BaseAPI/src/main/java/co/q64/arcade/base/api/util/TaskRunner.java
@@ -3,5 +3,5 @@ package co.q64.arcade.base.api.util;
 public interface TaskRunner {
 	public void runSync(Runnable task);
 
-	public void runSync(Runnable task, int ticks);
+	public void runSync(Runnable task, long ticks);
 }
